+++ StreamChatSwiftUITests/Tests/ChatChannel/ChannelInfo/MediaAttachmentsViewModel_Tests.swift
@@ -27,7 +27,7 @@ class MediaAttachmentsViewModel_Tests: StreamChatTestCase {
 
         // When
         let mediaItems = viewModel.mediaItems
-        let imageAttachments = viewModel.allImageAttachments
+        let imageAttachments = viewModel.allMediaAttachments.filter { $0.type == .image }
 
         // Then
         XCTAssert(mediaItems.count == 8)