+++ StreamChatSwiftUITests/Tests/ChatChannel/GalleryView_Tests.swift
@@ -48,8 +48,23 @@ class GalleryView_Tests: StreamChatTestCase {
 
     func test_gridView_snapshotLoading() {
         // Given
-        let view = GridPhotosView(
-            imageURLs: [ChatChannelTestHelpers.testURL],
+        let view = GridMediaView(
+            attachments: [MediaAttachment(url: ChatChannelTestHelpers.testURL, type: .image)],
+            isShown: .constant(true)
+        )
+        .applyDefaultSize()
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_gridViewVideoAndImage_snapshotLoading() {
+        // Given
+        let view = GridMediaView(
+            attachments: [
+                MediaAttachment(url: ChatChannelTestHelpers.testURL, type: .image),
+                MediaAttachment(url: ChatChannelTestHelpers.testURL.appendingPathComponent("test"), type: .video)
+            ],
             isShown: .constant(true)
         )
         .applyDefaultSize()